refactor: Make to Atomic design
