+++ src/components/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import useRtcClient from './hooks/useRtcClient';
-import InputForm from './InputForm';
-import VideoArea from './VideoArea';
+import InputForm from './templates/InputForm';
+import VideoArea from './templates/VideoArea';
 
 const App = () => {
   const rtcClient = useRtcClient();
