+++ src/components/organisms/InputFormLocal.tsx
@@ -13,8 +13,8 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 
-import Copyright from './organisms/Copyright';
-import RtcClient from '../utils/RtcClient';
+import Copyright from '../molecules/Copyright';
+import RtcClient from '../../utils/RtcClient';
 
 const useStyles = makeStyles((theme) => ({
   paper: {
@@ -38,7 +38,7 @@ type Props = {
 
 const InputFormLocal: VFC<Props> = (props) => {
   const { rtcClient } = props;
-  const label = 'remote peer name';
+  const label = 'your name';
   const classes = useStyles();
   const [disabled, setDisabled] = useState(true);
   const [name, setName] = useState('');
@@ -49,16 +49,15 @@ const InputFormLocal: VFC<Props> = (props) => {
     setDisabled(disabled);
   }, [name]);
 
-  const initializeRemotePeer = useCallback(
+  const initializeLocalPeer = useCallback(
     async (e: any) => {
-      await rtcClient.connect(name);
+      await rtcClient.startListening(name);
       e.preventDefault();
     },
     [name, rtcClient]
   );
 
-  if (rtcClient.localPeerName === '') return <></>;
-  if (rtcClient.remotePeerName !== '') return <></>;
+  if (rtcClient.localPeerName !== '') return <></>;
 
   return (
     <Container component="main" maxWidth="xs">
@@ -83,7 +82,7 @@ const InputFormLocal: VFC<Props> = (props) => {
               const target: any = e.target;
               if (target.value === '') return;
 
-              if (e.key === 'Enter') await initializeRemotePeer(e);
+              if (e.key === 'Enter') await initializeLocalPeer(e);
             }}
             required
             value={name}
@@ -94,7 +93,7 @@ const InputFormLocal: VFC<Props> = (props) => {
             color="primary"
             disabled={disabled}
             fullWidth
-            onClick={async (e) => await initializeRemotePeer(e)}
+            onClick={async (e) => await initializeLocalPeer(e)}
             type="submit"
             variant="contained"
           >
